refactor(extensions): split usage tracker module
