+++ .changeset/split-usage-tracker-into-provider-and-formatting-modules.md
@@ -1,0 +1,5 @@
+---
+default: patch
+---
+
+refactor the usage tracker by splitting shared types, formatting helpers, and provider probe logic into focused modules.
